requirements.txt
torch>=2.0.1
transformers>=4.30.2
streamlit>=1.22.0
nltk>=3.8.1
rake-nltk==1.0.6
PyPDF2>=3.0.1
python-docx>=0.8.11
sentencepiece>=0.1.99
